requirements.txt
langchain
openai
pinecone-client
langchain-pinecone
langchain-openai
python-dotenv
pypdf
flask
requests
